+++ unnaf/src/utils.c
@@ -1,9 +1,8 @@
 /*
  * NAF decompressor
- * Copyright (c) 2018-2020 Kirill Kryukov
+ * Copyright (c) 2018-2021 Kirill Kryukov
  * See README.md and LICENSE files of this repository
  */
-
 
 //__attribute__ ((format (printf, 1, 2)))
 static void msg(const char *format, ...) 
@@ -39,6 +38,13 @@ static void die(const char *format, ...)
     va_end(argptr);
     exit(1);
 }
+
+
+#define ZSTD_TRY(f)  \
+do {                 \
+    size_t e = f;    \
+    if (ZSTD_isError(e)) { die("zstd error: %s", ZSTD_getErrorName(e)); }  \
+} while (0)
 
 
 __attribute__ ((cold))